+++ include/Core/SPK_Octree.h
@@ -33,7 +33,7 @@ namespace SPK
 	* When a octree is attached to a group, at each update, every individual particle is processed and put into cells.<br>
 	* Particles become therefore aware of their neighbors.<br>
 	* <br>
-	* Octrees allows optimization of algorithms that run in O(n²) by reducing their complexity to O(nlog(n)).<br>
+	* Octrees allows optimization of algorithms that run in O(n? by reducing their complexity to O(nlog(n)).<br>
 	* Typically algorithms where each particle is affected by every other particles in the group (particle vs particle collision, flocking, nbody simulations...).<br>
 	* <br>
 	* A Octree is automatically generated within a group if at least one of its modifiers needs it (by setting its NEEDS_OCTREE constant to true at init).<br>
@@ -58,11 +58,11 @@ namespace SPK
 
 			~Array<T>() { SPK_DELETE_ARRAY(values); }
 
-			size_t size() const					{ return currentNb; }
-			size_t capacity() const				{ return maxNb; }
+			unsigned int size() const					{ return currentNb; }
+			unsigned int capacity() const				{ return maxNb; }
 			bool empty() const					{ return currentNb == 0; }
-			T& operator[](size_t i)				{ return values[i]; }
-			const T& operator[](size_t i) const	{ return values[i]; }
+			T& operator[](unsigned int i)				{ return values[i]; }
+			const T& operator[](unsigned int i) const	{ return values[i]; }
 
 			void push(T t)
 			{
@@ -71,7 +71,7 @@ namespace SPK
 					// Reallocates
 					maxNb <<= 1;
 					T* tmp = SPK_NEW_ARRAY(T,maxNb);
-					for (size_t i = 0; i < currentNb; ++i)
+					for (unsigned int i = 0; i < currentNb; ++i)
 						tmp[i] = values[i];
 					SPK_DELETE_ARRAY(values);
 					values = tmp;
@@ -84,11 +84,11 @@ namespace SPK
 
 		/*private*/public : // tmp workaround because I wasnt able to allows Cell to be friend of Array (compiler bug ?) but thats unsafe
 
-			size_t currentNb;
-			size_t maxNb;
+			unsigned int currentNb;
+			unsigned int maxNb;
 			T* values;
 
-			Array<T>(size_t maxNb = 1) :
+			Array<T>(unsigned int maxNb = 1) :
 				currentNb(0),
 				maxNb(maxNb)
 			{
@@ -100,7 +100,7 @@ namespace SPK
 				maxNb(t.maxNb)
 			{
 				values = SPK_NEW_ARRAY(T,maxNb);
-				for (size_t i = 0; i < currentNb; ++i)
+				for (unsigned int i = 0; i < currentNb; ++i)
 					values[i] = t.values[i];
 			} // Not called
 
@@ -114,7 +114,7 @@ namespace SPK
 					SPK_DELETE_ARRAY(values);
 					values = SPK_NEW_ARRAY(T,maxNb);
 				}
-				for (size_t i = 0; i < currentNb; ++i)
+				for (unsigned int i = 0; i < currentNb; ++i)
 					values[i] = t.values[i];
 				return *this;
 			}
@@ -128,17 +128,17 @@ namespace SPK
 
 		public :
 
-			size_t level;
-			size_t offsetX;
-			size_t offsetY;
-			size_t offsetZ;
-			size_t children[8];
+			unsigned int level;
+			unsigned int offsetX;
+			unsigned int offsetY;
+			unsigned int offsetZ;
+			unsigned int children[8];
 			bool hasChildren;
-			Array<size_t> particles;
+			Array<unsigned int> particles;
 
 		private :
 
-			void init(size_t level,size_t offsetX,size_t offsetY,size_t offsetZ)
+			void init(unsigned int level,unsigned int offsetX,unsigned int offsetY,unsigned int offsetZ)
 			{
 				this->level = level;
 				this->offsetX = offsetX;
@@ -148,7 +148,7 @@ namespace SPK
 				particles.clear();
 			}
 
-			Cell(size_t level = 0,size_t offsetX = 0,size_t offsetY = 0,size_t offsetZ = 0) :
+			Cell(unsigned int level = 0,unsigned int offsetX = 0,unsigned int offsetY = 0,unsigned int offsetZ = 0) :
 				level(level),
 				offsetX(offsetX),
 				offsetY(offsetY),
@@ -175,7 +175,7 @@ namespace SPK
 				offsetY = cell.offsetY;
 				offsetZ = cell.offsetZ;
 				hasChildren = cell.hasChildren;
-				std::memcpy(children,cell.children,8 * sizeof(size_t));
+				std::memcpy(children,cell.children,8 * sizeof(unsigned int));
 				particles = cell.particles;
 				return *this;
 			}
@@ -186,7 +186,7 @@ namespace SPK
 		* Active cells are cells of the octree that contains at least one particle
 		* @return active cells of the octree
 		*/
-		const Array<size_t>& getActiveCells() const							{ return activeCells; }
+		const Array<unsigned int>& getActiveCells() const							{ return activeCells; }
 
 		/**
 		* @brief Gets the neighboring cells of a given particle
@@ -196,14 +196,14 @@ namespace SPK
 		* @param particleIndex ; the index of the particle
 		* @return the neighboring cells of a given particle
 		*/
-		const Array<size_t>& getNeighborCells(size_t particleIndex) const	{ return particleCells[particleIndex]; }
+		const Array<unsigned int>& getNeighborCells(unsigned int particleIndex) const	{ return particleCells[particleIndex]; }
 
 		/**
 		* @brief Gets a given cell by index
 		* @param index : the index of the cell
 		* @return the cell at the given index
 		*/
-		const Cell& getCell(size_t index) const								{ return cells[index]; }
+		const Cell& getCell(unsigned int index) const								{ return cells[index]; }
 
 		/**
 		* @brief Gets the minimum position of the axis aligned bounding box of the octree
@@ -233,20 +233,20 @@ namespace SPK
 			}
 		};
 
-		static const size_t MAX_LEVEL_INDEX;
-		static const size_t MAX_PARTICLES_NB_PER_CELL;
+		static const unsigned int MAX_LEVEL_INDEX;
+		static const unsigned int MAX_PARTICLES_NB_PER_CELL;
 		static const float OPTIMAL_CELL_SIZE_FACTOR;
 		static const float MIN_CELL_SIZE;
 
 		Group& group;
 
 		Array<Cell> cells; // Pool of cells
-		size_t nbCells;
+		unsigned int nbCells;
 
-		Array<size_t> activeCells; // Index of the active cells
+		Array<unsigned int> activeCells; // Index of the active cells
 
-		Array<size_t>* particleCells; // Cells to which particles belongs
-		size_t nbParticles;
+		Array<unsigned int>* particleCells; // Cells to which particles belongs
+		unsigned int nbParticles;
 
 		Triplet* minPos;
 		Triplet* maxPos;
@@ -263,9 +263,9 @@ namespace SPK
 
 		void update();  // Used by Group only
 
-		size_t initNextCell(size_t level,size_t offsetX,size_t offsetY,size_t offsetZ);
-		void addToCell(size_t cellIndex,size_t particleIndex,size_t maxLevel);
-		void addToChildrenCells(size_t parentIndex,size_t particleIndex,size_t maxLevel);
+		unsigned int initNextCell(unsigned int level,unsigned int offsetX,unsigned int offsetY,unsigned int offsetZ);
+		void addToCell(unsigned int cellIndex,unsigned int particleIndex,unsigned int maxLevel);
+		void addToChildrenCells(unsigned int parentIndex,unsigned int particleIndex,unsigned int maxLevel);
 	};
 }
