+++ include/Core/SPK_System.h
@@ -31,13 +31,13 @@ private : \
 SPK::Ref<SPK::System> SPKSystem; \
 public : \
 const SPK::Ref<SPK::System>& getSPKSystem() { return SPKSystem; } \
-SPK::Ref<SPK::Group> createSPKGroup(size_t capacity) { return SPKSystem->createGroup(capacity); } \
+SPK::Ref<SPK::Group> createSPKGroup(unsigned int capacity) { return SPKSystem->createGroup(capacity); } \
 SPK::Ref<SPK::Group> createSPKGroup(SPK::Ref<SPK::Group>& group) { return SPKSystem->createGroup(group); } \
 void addSPKGroup(SPK::Ref<SPK::Group>& group) { SPKSystem->addGroup(group); } \
 void removeSPKGroup(SPK::Ref<SPK::Group> group) { SPKSystem->removeGroup(group); } \
-SPK::Ref<SPK::Group> getSPKGroup(size_t index) const { return SPKSystem->getGroup(index); } \
-size_t getNbSPKGroups() const { return SPKSystem->getNbGroups(); } \
-size_t getNbParticles() const { return SPKSystem->getNbParticles(); } \
+SPK::Ref<SPK::Group> getSPKGroup(unsigned int index) const { return SPKSystem->getGroup(index); } \
+unsigned int getNbSPKGroups() const { return SPKSystem->getNbGroups(); } \
+unsigned int getNbParticles() const { return SPKSystem->getNbParticles(); } \
 void initializeSPK() { SPKSystem->initialize(); } \
 bool isInitializedSPK() const { return SPKSystem->isInitialized(); }
 
@@ -96,7 +96,7 @@ namespace SPK
 		/**
 		* @brief Returns the number of controllers in the system
 		*/
-		size_t getNbControllers() const;
+        size_t getNbControllers() const;
 
 		///////////////////////
 		// Groups management //
@@ -106,7 +106,7 @@ namespace SPK
 		* @brief Adds a group to the system
 		* @param group : a pointer on the group to add to the system
 		*/
-		Ref<Group> createGroup(size_t capacity);
+		Ref<Group> createGroup(unsigned int capacity);
 
 		/**
 		* @brief Adds a group to the system which is a copy of an existing group
@@ -141,13 +141,13 @@ namespace SPK
 		* @brief Gets the number of groups in the system
 		* @return the number of groups in the system
 		*/
-		size_t getNbGroups() const;
+        size_t getNbGroups() const;
 
 		/**
 		* @brief Gets the number of particles in the system
 		* @return the number of particles in the system
 		*/
-		size_t getNbParticles() const;
+		unsigned int getNbParticles() const;
 
 		/////////////////////////////
 		// Operations on particles //
@@ -395,7 +395,7 @@ namespace SPK
 
 	inline size_t System::getNbControllers() const
 	{
-		return controllers.size();
+		return static_cast<unsigned int>(controllers.size());
 	}
 
 	inline void System::enableAABBComputation(bool AABB)
